Final code of Country Maximum Normalisation
